Minor clean up to the code and update READEME.
When copying over the original code to the new repository, whitespace was removed. This changes adds that back in.

It also updates the README to include limitations and removes Google from the tool name as per the open sourcing checklist.

Finally, it changes the IDs used in the tests, to make it clearer that these are not real account numbers.

Change-Id: I2b9e397cf6e513742dd15b70e0497575acbdb28b